requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
black==26.1.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
Cython==3.2.4
decorator==5.2.1
Django==6.0.3
django-debug-toolbar==6.2.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
ffmpeg-python==0.2.0
filelock==3.25.0
ftfy==6.3.1
future==1.0.0
gallery_dl==1.31.7
humanize==4.15.0
idna==3.11
inflection==0.5.1
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
library==3.0.167
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mutagen==1.47.0
mypy_extensions==1.1.0
natsort==8.4.0
packaging==26.0
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
platformdirs==4.9.2
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==2.0.0
pycryptodomex==3.23.0
Pygments==2.19.2
PyJWT==2.11.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-mpv-jsonipc==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
rpds-py==0.30.0
screeninfo==0.8.1
six==1.17.0
soupsieve==2.8.3
sqlite-fts4==1.0.3
sqlite-utils==3.39
sqlparse==0.5.5
stack-data==0.6.3
tabulate==0.10.0
tinytag==2.2.0
tldextract==5.3.1
traitlets==5.14.3
typing_extensions==4.15.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
yt-dlp==2026.3.3
yt-dlp-ejs==0.5.0
